pyproject.toml
[project]
name = "chat-a-doc"
version = "0.8.1"
description = "MCP server for generating documents in multiple formats (PDF, DOCX, HTML, TXT, CSV, MARKDOWN) from markdown content with Word template support."
requires-python = ">=3.11,<3.14"
dependencies = [
 "mcp>=1.2.1",
 "pyyaml>=6.0.2",
 "markdown>=3.5",  # For HTML, TXT, CSV, PDF, DOCX generation
 "html2text>=2024.2.26",  # For TXT generation
 "weasyprint>=60",  # For PDF generation
 "python-docx>=1.1",  # For DOCX generation
]
[[project.authors]]
name = "Vivek Vellaiyappan Surulimuthu"
email = "vivekvellaiyappans@gmail.com"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chat_a_doc"]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.0.0",
    "pytest>=8.4.1",
    "ruff>=0.12.11",
    "jsonschema>=4.25.1",
    "yamllint>=1.37.1",
    "pre-commit>=4.3.0",
]

[project.scripts]
chat-a-doc = "chat_a_doc:main"
[tool.ruff]
line-length = 120
exclude = ["tests/*"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "D", "N", "S"]
ignore = []

[tool.yamllint]
config-file = ".yamllint.yaml"
